created jsonhandler.go
